Add single pass API
